pyproject.toml
[project]
name = "automation-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "behave>=1.3.0",
    "jsonschema>=4.25.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "selenium>=4.34.2",
    "webdriver-manager>=4.0.2",
]
